Split the build and push in two separate opposing steps
We build and use the caches when we are just running this job in a PR
(i.e. we are not pushing anything to quay and we just want to make sure
it can be built correctly)

When we do an actual push we stop using cached layers so we make sure we
pull in the very latest git repos for cluster_utils.